\name{mbqc-package}
\alias{mbqc-package}
\alias{mbqc}
\docType{package}
\title{
\packageTitle{mbqc}
}
\description{
\packageDescription{mbqc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mbqc}
\packageIndices{mbqc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mbqc}

Maintainer: \packageMaintainer{mbqc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
